+++ src/main/scala/view/plague/TraitsView.scala
@@ -1,7 +1,8 @@
 package view.plague
 
 import model.core.SimulationState
-import model.plague.{Symptoms, Trait}
+import model.plague.Trait
+import model.plague.db.Symptoms
 import scalafx.geometry.Insets
 import scalafx.scene.layout.BorderPane
 import scalafx.Includes.jfxReadOnlyObjectProperty2sfx
